perf(dashboard): move static chrome to Astro shells (Settings, Argo table)
Settings: split the 782-line ReactSettingsPanel into a static SettingsCard.astro
shell (header chrome + inline SVG icons, server HTML, zero hydration) plus three
focused body islands (Storage / Health / Device). The System Health count badge
is driven imperatively from the island via a closest()-scoped data-slot write —
no vDOM for the card frame.

Argo: lift the runtime-injected ARGO_TABLE_CSS (re-emitted on every island
render via <style>{...}</style>) into AstroArgoDashboard global CSS, shipping the
grid layout server-side and dropping the style element + string from the island.
